[GAIA/engine] Benchmark: merge k-hop project into benchmark project;
