+++ research/engine/pegasus/benchmark/Dockerfile
@@ -2,7 +2,7 @@ FROM rust:1.55.0
 WORKDIR /usr/src/pegasus/
 COPY .. .
 RUN rm -f Cargo.lock && \
-    cd ./khop && rm -f Cargo.lock && \
+    cd ./benchmark && rm -f Cargo.lock && \
     cd .. && cp ./crates-io.config /usr/local/cargo/config && \
-    cd ./khop && cargo install --path . && rm -rf target
-CMD ["pegasus-khop"]
+    cd ./benchmark && cargo install --path . && rm -rf target
+CMD ["service"]
