+++ research/engine/pegasus/benchmark/src/bin/spmsk.rs
@@ -1,29 +1,32 @@
 use std::path::PathBuf;
 use std::sync::Arc;
 use std::time::Instant;
-
-use pegasus::api::{Count, Iteration, Map, Sink};
 use pegasus::{Configuration, JobConf, ServerConf};
 use structopt::StructOpt;
 
+/// Searching and counting k-hop neighbors for each vertex in a list use only one job;
+/// Do k-hop searching for all vertices in barrier sync mode;
 #[derive(Debug, StructOpt)]
-#[structopt(name = "khop", about = "Search khop neighbors on parallel dataflow")]
+#[structopt(name = "sync packed multi-src k-hop", about = "Search k-hop neighbors using parallel dataflow system")]
 struct Config {
     /// The number of hop this job will search;
     #[structopt(short = "k", default_value = "3")]
     k: u32,
-    #[structopt(short = "n", default_value = "100")]
+    #[structopt(short = "t", default_value = "100")]
     starts: u32,
-    #[structopt(short = "i")]
-    id_from_std: bool,
     #[structopt(short = "f")]
     use_loop: bool,
+    /// specify the ids of start vertices;
+    #[structopt(short = "i")]
+    id_from_std: bool,
     /// The path of the origin graph data ;
     #[structopt(long = "data", parse(from_os_str))]
     data_path: PathBuf,
     /// the number of partitions to partition the local graph;
     #[structopt(short = "p", default_value = "1")]
     partitions: u32,
+    #[structopt(short = "b", default_value = "64")]
+    batch_width: u32,
     #[structopt(short = "s", long = "servers")]
     servers: Option<PathBuf>,
 }
@@ -39,8 +42,10 @@ fn main() {
     };
     pegasus::startup(server_conf).unwrap();
     let graph = Arc::new(pegasus_graph::load(&config.data_path).unwrap());
-    let mut conf = JobConf::new("k_hop");
+    // config job;
+    let mut conf = JobConf::new("de_correlated_k_hop");
     conf.set_workers(config.partitions);
+    conf.batch_capacity = config.batch_width;
     if config.servers.is_some() {
         conf.reset_servers(ServerConf::All);
     }
@@ -69,75 +74,34 @@ fn main() {
 
     pegasus::wait_servers_ready(conf.servers());
 
-    let mut results = Vec::new();
     let k_hop = config.k;
     let nums = src.len();
-
     if nums == 0 {
         return;
     }
 
     println!("start search {}-hop neighbors for {} vertices;", k_hop, nums);
 
+    assert!(k_hop > 0);
     let use_loop = config.use_loop;
-    let global_start = Instant::now();
-    for (i, id) in src.into_iter().enumerate() {
-        let mut conf = conf.clone();
-        conf.job_id = i as u64;
-        conf.plan_print = i <= 0;
-        let start = Instant::now();
-        let result = pegasus::run(conf.clone(), || {
-            let index = pegasus::get_current_worker().index;
-            let graph = graph.clone();
-            let src = if index == 0 { vec![id] } else { vec![] };
-            move |input, output| {
-                let mut stream = input.input_from(src)?;
-                if use_loop {
-                    stream = stream.iterate(k_hop, |start| {
-                        let graph = graph.clone();
-                        start
-                            .repartition(|id| Ok(*id))
-                            .flat_map(move |id| Ok(graph.get_neighbors(id)))
-                    })?;
-                } else {
-                    for _i in 0..k_hop {
-                        let graph = graph.clone();
-                        stream = stream
-                            .repartition(|id| Ok(*id))
-                            .flat_map(move |id| Ok(graph.get_neighbors(id)))?;
-                    }
-                }
-                stream
-                    .count()?
-                    .map(move |cnt| {
-                        let x = start.elapsed();
-                        Ok((id, cnt, x))
-                    })?
-                    .sink_into(output)
-            }
-        })
-        .expect("submit job failure;");
-        results.push(result);
-    }
+    let start = Instant::now();
+    let res = pegasus_benchmark::queries::khop::packed_multi_src_k_hop(src, k_hop, use_loop, conf, graph);
 
     let mut cnt_list = Vec::new();
     let mut elp_list = Vec::new();
     let mut n = 10;
 
-    for res in results {
-        for x in res {
-            let (id, cnt, elp) = x.unwrap();
-            if cnt > 0 {
-                if n > 0 {
-                    println!("{}\tfind khop\t{}\tneighbors, use {:?};", id, cnt, elp);
-                }
-                n -= 1;
-                cnt_list.push(cnt);
-                elp_list.push(elp.as_millis() as u64);
-            }
+    for x in res {
+        let (id, cnt, elp) = x.unwrap();
+        if n > 0 {
+            println!("{}\tfind khop\t{}\tneighbors, use {:?}ms;", id, cnt, elp);
         }
+        n -= 1;
+        cnt_list.push(cnt);
+        elp_list.push(elp);
     }
-    let elp = global_start.elapsed();
+
+    let elp = start.elapsed();
     println!("...");
     println!("==========================================================");
 
@@ -169,7 +133,8 @@ fn main() {
     println!("avg elapse {} ms", total as f64 / len);
 
     println!("==========================================================");
-    let millis = elp.as_millis() as f64;
-    println!("total use {}ms, qps: {:.1}", millis, nums as f64 * 1000.0 / millis);
+    let millis = elp.as_millis() as u64;
+    println!("total use {}ms, qps: {:.1}", millis, nums as f64 * 1000f64 / millis as f64);
+
     pegasus::shutdown_all();
 }
