+++ research/engine/pegasus/benchmark/src/queries/khop/multi_src/packed.rs
@@ -8,7 +8,7 @@ use pegasus::tag::tools::map::TidyTagMap;
 use pegasus::JobConf;
 
 use super::super::one_hop;
-use crate::Graph;
+use crate::graph::Graph;
 
 pub fn packed_multi_src_k_hop<G: Graph>(
     src: Vec<u64>, k_hop: u32, use_loop: bool, conf: JobConf, graph: Arc<G>,
