--- research/engine/pegasus/khop/src/main.rs
@@ -1,3 +1,0 @@
-fn main() {
-    println!("This is k-hop implementation by pegasus;");
-}
